added exceptions and response codes

+++ location/src/main/java/com/example/location/controller/LocationController.java
@@ -1,6 +1,9 @@
 package com.example.location.controller;
 
+import javax.validation.Valid;
+
 import com.example.location.entity.Location;
+import com.example.location.errors.LocationNotFoundException;
 import com.example.location.service.LocationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -16,6 +19,9 @@ public class LocationController {
 
     @GetMapping("/{id}")
     public Location FindLocationById(@PathVariable("id") Long locationId){
+        if (locationService.findLocationById(locationId) == null){
+            throw new LocationNotFoundException("Location not found");
+        }
         return locationService.findLocationById(locationId);
     }
 
@@ -25,17 +31,21 @@ public class LocationController {
     }
 
     @PostMapping("/add")
-    public Location saveLocation(@RequestBody Location newLocation) {
+    public Location saveLocation(@Valid @RequestBody Location newLocation) {
         return locationService.saveLocation(newLocation);
     }
 
     @PutMapping("/edit/{id}")
-    public Location updateLocation(@RequestBody Location updateLocation, @PathVariable Long id) {
+    public Location updateLocation(@Valid @RequestBody Location updateLocation, @PathVariable Long id) {
         return locationService.updateLocation(updateLocation, id);
     }
 
     @DeleteMapping("/delete/{id}")
     public void deleteLocation(@PathVariable Long id) {
+        if (locationService.findLocationById(id) == null){
+            throw new LocationNotFoundException("Location not found");
+        }
+
         locationService.deleteById(id);
     }
 }
